+++ areas/monitor/src/AzureMcp.Monitor/MonitorSetup.cs
@@ -8,6 +8,7 @@ using AzureMcp.Monitor.Commands.Log;
 using AzureMcp.Monitor.Commands.Metrics;
 using AzureMcp.Monitor.Commands.Table;
 using AzureMcp.Monitor.Commands.TableType;
+using AzureMcp.Monitor.Commands.WebTests;
 using AzureMcp.Monitor.Commands.Workspace;
 using AzureMcp.Monitor.Services;
 using Microsoft.Extensions.DependencyInjection;
@@ -21,6 +22,7 @@ public class MonitorSetup : IAreaSetup
     {
         services.AddSingleton<IMonitorService, MonitorService>();
         services.AddSingleton<IMonitorHealthModelService, MonitorHealthModelService>();
+        services.AddSingleton<IMonitorWebTestService, MonitorWebTestService>();
         services.AddSingleton<IResourceResolverService, ResourceResolverService>();
         services.AddSingleton<IMetricsQueryClientService, MetricsQueryClientService>();
         services.AddSingleton<IMonitorMetricsService, MonitorMetricsService>();
@@ -75,5 +77,13 @@ public class MonitorSetup : IAreaSetup
 
         metrics.AddCommand("query", new MetricsQueryCommand(loggerFactory.CreateLogger<MetricsQueryCommand>()));
         metrics.AddCommand("definitions", new MetricsDefinitionsCommand(loggerFactory.CreateLogger<MetricsDefinitionsCommand>()));
+
+        // Register Monitor.WebTest sub-group commands
+        var webTests = new CommandGroup("webtests", "Azure Monitor Web Test operations - Commands for working with Azure Availability/Web Tests.");
+        monitor.AddSubGroup(webTests);
+
+        webTests.AddCommand("get", new WebTestsGetCommand(loggerFactory.CreateLogger<WebTestsGetCommand>()));
+        webTests.AddCommand("list", new WebTestsListCommand(loggerFactory.CreateLogger<WebTestsListCommand>()));
+        webTests.AddCommand("createorupdate", new WebTestsCreateOrUpdateCommand(loggerFactory.CreateLogger<WebTestsCreateOrUpdateCommand>()));
     }
 }