+++ areas/monitor/src/AzureMcp.Monitor/Options/AppInsights/BaseAppInsightsOptions.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using AzureMcp.Monitor.Options;
+
+namespace AzureMcp.Monitor.Options.AppInsights;
+
+public class BaseAppInsightsOptions : BaseMonitorOptions;