+++ areas/monitor/src/AzureMcp.Monitor/Options/AppInsights/WebTests/WebTestsGetOptions.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+
+namespace AzureMcp.Monitor.Options.AppInsights.WebTests;
+
+public class WebTestsGetOptions : BaseAppInsightsOptions
+{
+    [JsonPropertyName(MonitorOptionDefinitions.WebTestResourceName)]
+    public string? ResourceName { get; set; }
+}