+++ areas/monitor/src/AzureMcp.Monitor/Options/AppInsights/WebTests/WebTestsListOptions.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Monitor.Options.AppInsights.WebTests;
+
+public class WebTestsListOptions : BaseAppInsightsOptions
+{
+}